+++ Sources/AppIconKit/AlternateAppIconShelf.swift
@@ -29,12 +29,14 @@ public struct AlternateAppIconShelf: View {
         context: AlternateAppIconContext,
         onIconSelected: @escaping (AlternateAppIcon) -> Void = { _ in }
     ) {
-        self.collections = collections
+        self.sections = collections.map { .init(title: $0.name, items: $0.icons) }
         self.context = context
         self.onIconSelected = onIconSelected
     }
+    
+    public typealias Section = ItemShelfSection<AlternateAppIcon>
 
-    private let collections: [AlternateAppIconCollection]
+    private let sections: [Section]
     private let onIconSelected: (AlternateAppIcon) -> Void
 
     @Environment(\.alternateAppIconShelfStyle)
@@ -44,49 +46,13 @@ public struct AlternateAppIconShelf: View {
     private var context: AlternateAppIconContext
 
     public var body: some View {
-        ScrollView(.vertical) {
-            LazyVStack(spacing: style.sectionSpacing) {
-                ForEach(Array(collections.enumerated()), id: \.offset) { section in
-                    VStack(alignment: .leading, spacing: style.sectionTitleSpacing) {
-                        shelfTitle(for: section.element)
-                        shelf(for: section.element)
-                    }
-                }
-            }
-            .padding(.vertical, style.scrollPadding)
-        }
-        .withHiddenScrollContent()
-        .background(style.backgroundColor)
-        .alternateAppIconListItemStyle(style.itemStyle)
-    }
-}
-
-private extension View {
-
-    func withHiddenScrollContent() -> some View {
-        #if os(tvOS)
-        self
-        #else
-        self.scrollContentBackground(.hidden)
-        #endif
+        ItemShelf(sections: sections, itemView: itemView)
     }
 }
 
 private extension AlternateAppIconShelf {
 
-    func shelf(for collection: AlternateAppIconCollection) -> some View {
-        ScrollView(.horizontal, showsIndicators: false) {
-            LazyHStack(spacing: style.itemSpacing) {
-                ForEach(Array(collection.icons.enumerated()), id: \.offset) {
-                    shelfItem(for: $0.element)
-                }
-            }
-            .padding(.horizontal, style.sectionPadding)
-        }
-        .scrollClipDisabled()
-    }
-
-    func shelfItem(for icon: AlternateAppIcon) -> some View {
+    func itemView(for icon: AlternateAppIcon) -> some View {
         Button {
             selectIcon(icon)
         } label: {
@@ -96,15 +62,6 @@ private extension AlternateAppIconShelf {
             )
         }
         .buttonStyle(.plain)
-    }
-
-    func shelfTitle(for collection: AlternateAppIconCollection) -> some View {
-        Text(collection.name)
-            .font(.footnote)
-            .textCase(.uppercase)
-            .foregroundColor(.secondary)
-            .padding(.horizontal, style.sectionPadding)
-            .padding(.horizontal, 0.1 * style.itemStyle.iconSize)
     }
 }
 
@@ -121,50 +78,7 @@ private extension AlternateAppIconShelf {
 public extension AlternateAppIconShelf {
 
     /// This style can style a ``AlternateAppIconShelf``.
-    struct Style: Sendable {
-
-        /// Create a custom style.
-        ///
-        /// - Parameters:
-        ///   - scrollPadding: The vertical scroll padding, by default `20`.
-        ///   - sectionSpacing: The spacing between sections, by default `40`.
-        ///   - sectionTitleSpacing: The spacing between section title and items, by default `10`.
-        ///   - sectionPadding: The horizontal padding of each section, by default `nil`.
-        ///   - itemStyle: The item style, by default `.standard`.
-        ///   - itemSpacing: The spacing between items, by default `16`.
-        ///   - backgroundColor: The shelf background color.
-        public init(
-            scrollPadding: Double = 20,
-            sectionSpacing: Double = 40,
-            sectionTitleSpacing: Double = 10,
-            sectionPadding: CGFloat? = nil,
-            itemStyle: AlternateAppIconListItem.Style = .standard,
-            itemSpacing: Double = 16,
-            backgroundColor: Color = .primary.opacity(0.05)
-        ) {
-            self.scrollPadding = scrollPadding
-            self.sectionSpacing = sectionSpacing
-            self.sectionTitleSpacing = sectionTitleSpacing
-            self.sectionPadding = sectionPadding
-            self.itemStyle = itemStyle
-            self.itemSpacing = itemSpacing
-            self.backgroundColor = backgroundColor
-        }
-
-        public var scrollPadding: Double
-        public var sectionSpacing: Double
-        public var sectionTitleSpacing: Double
-        public var sectionPadding: CGFloat?
-        public var itemStyle: AlternateAppIconListItem.Style
-        public var itemSpacing: Double
-        public var backgroundColor: Color
-    }
-}
-
-public extension AlternateAppIconShelf.Style {
-
-    /// The standard alternate app icon item style.
-    static var standard: Self { .init() }
+    typealias Style = ItemShelfStyle
 }
 
 public extension View {